This unobtrusive JavaScript module finds all <a> elements ...
This unobtrusive JavaScript module finds all <a> elements that have an href attribute but no title attribute and adds an onmouseover event handler to them. The event handler makes an XMLHttpRequest HEAD request to fetch details about the linked resource, and then sets those details in the title attribute of the link so that they will be displayed as a tooltip.