Return a newly created <img> element that will be set to display a Google map of the current location
Return a newly created <img> element that will (once geolocation succeeds) be set to display a Google map of the current location. Note that the caller must insert the returned element into the document in order to make it visible. Throws an error if geolocation is not supported in the browser